Evas Textblock: move line breaks getting to function
Calculating line breaks moved to an outside function since its usage
does not really go well with the nice structure of _layout_par.
Also, moved the line in which we set the line breaks a bit further to
the part that actually uses that, and added a small comment about its
usage.