Evas Textblock: Modify layout to handle dirty parts
This introduces faster handling for cases where text is added to an
already laid-out paragraph. The original handling was to clear all the
paragraph's content and relayout from scratch, upon every change.
Now, when a text is added, the affected Text Items are tested and
rebuilt. This saves a significant amount of pre-layout time, especially with
actions which require fast responses e.g. typing in an Entry widget.