evas textblock: Add Obstacles support
Introducing a new feature for Evas Textblock. This allows the layout to
wrap around other evas objects.
The following API is added (eo):
 - obstacle_add
 - obstacle_del
 - obstacles_update
Using this is pretty straight-forward: you add the evas object via
obstacle_add, and flag for a relayout when ready with obstacles_update.
This is passive and does not affect the registered objects, and also
there is no need for ref'ing.
There is also an example to get a grasp of this feature.

@feature