evas/textblock: add obstacles feature
The obstacle features allows registering evas objects to affect the
layout of the text. The text is arranged around an obstacle's geometry
so none of its content is hidden.
Several modes (formats) are currently supported: FLOAT (wrap around),
LINE (wrap as if the obstacle is several lines of text) and DISABLED,
of course.
See the docs for full info.

@feature